Extract the receipt contract into a composable
